refactor: 프로젝트 수정 로직 분리 (#174)
