+++ src/components/domain/project/projectButton/ProjectEditDeleteButton.tsx
@@ -1,9 +1,6 @@
 import { Edit3, Trash2 } from 'lucide-react';
 
-import { useModalStore } from '@/stores/modalStore';
-import { useGetProjectDetails } from '@/hooks/queries/useProjectService';
-import ProjectRegisterModal from '../projectRegisterModal/ProjectRegisterModal';
-import { ProjectForm } from '@/models/project/projectModels';
+import useProjectUpdateModal from '../hooks/useProjectUpdateModal';
 import useConfirmDeleteProject from '../hooks/useConfirmDeleteProject';
 
 interface ProjectEditDeleteButtonProps {
@@ -11,29 +8,17 @@ interface ProjectEditDeleteButtonProps {
 }
 
 export default function ProjectEditDeleteButton({ projectId }: ProjectEditDeleteButtonProps) {
-  const { openModal } = useModalStore();
-
-  const handleGetDetailSuccess = (projectDetailData: ProjectForm) => {
-    openModal(
-      <ProjectRegisterModal isEdit projectId={projectId} defaultData={projectDetailData} />,
-    );
-  };
-  const getDetailMutation = useGetProjectDetails(handleGetDetailSuccess);
-
+  const { handleOpenProjectUpdateModal, isDetailLoading } =
+    useProjectUpdateModal<HTMLButtonElement>(projectId);
   const { handleConfirmDeleteProject } = useConfirmDeleteProject<HTMLButtonElement>(projectId);
-
-  const handleEditProject = (event: React.MouseEvent<HTMLButtonElement>) => {
-    event.stopPropagation();
-    getDetailMutation.mutate(projectId);
-  };
 
   return (
     <nav className="flex gap-3">
+      <button aria-label="편집" disabled={isDetailLoading} onClick={handleOpenProjectUpdateModal}>
+        <Edit3 className="stroke-gray-600 h-6 w-6 stroke-[1.5px] hover:stroke-gray-700 hover:stroke-[2px]" />
+      </button>
       <button aria-label="삭제" onClick={handleConfirmDeleteProject}>
         <Trash2 className="stroke-gray-600 h-6 w-6 stroke-[1.5px] hover:stroke-gray-700 hover:stroke-[2px]" />
-      </button>
-      <button aria-label="편집" disabled={getDetailMutation.isPending} onClick={handleEditProject}>
-        <Edit3 className="stroke-gray-600 h-6 w-6 stroke-[1.5px] hover:stroke-gray-700 hover:stroke-[2px]" />
       </button>
     </nav>
   );
